+++ src/hooks/useMovieTrailer.js
@@ -1,0 +1,30 @@
+import { useDispatch } from "react-redux";
+import { API_OPTIONS } from "../utils/constants";
+import { addTrailerVideo } from "../utils/moviesSlice";
+import { useEffect } from "react";
+
+const useMovieTrailer = ({ movieId }) => {
+  // fetch trailer video and updating the store with trailer video
+
+  const dispatch = useDispatch();
+  const getMovieVideos = async () => {
+    const data = await fetch(
+      "https://api.themoviedb.org/3/movie/" +
+        movieId +
+        "/videos?language=en-US",
+      API_OPTIONS
+    );
+    const json = await data.json();
+
+    const filterTrailer = json.results.filter(
+      (video) => video.type === "Trailer"
+    );
+    const trailer = filterTrailer.length ? filterTrailer[0] : json.results[0];
+    dispatch(addTrailerVideo(trailer));
+  };
+  useEffect(() => {
+    getMovieVideos();
+  }, []);
+};
+
+export default useMovieTrailer;
